requirements.txt
sphinx
sphinx-rtd-theme
selenium
pytest
webdriver-manager
flask
flask-bootstrap4
pyotp